pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learned-image-compression-ae"
version = "0.1.0"
description = "Autoencoder-based learned lossy image compression with adaptive importance maps."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "pymongo>=4.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
